fix ,upd runjar
1)
..config:
.. .gitignore:
..   fix DIR:target
..   added for eclipseIDE
..   added regex for log file [keylogger-*-*-*.txt]
.. pom.xml:
..   created variable for dependency version
..   created variable for main class
..   updated dependency
..   added ability to build executable jar file [build->plugins->maven-assembly-plugin]

2)
removed target dir

3)
.. KeyLogger.java:
..   push data to file on every event [saveLog()-> bw.flush();]
