+++ .gitignore
@@ -1,0 +1,7 @@
+.classpath
+.project
+.settings
+
+target
+
+keylogger-*-*-*.txt
